fish and links
add a generation of fishes
add a linking between objects

+++ src/Game/Object/Boat/Boat.cpp
@@ -1,6 +1,6 @@
 #include "Boat.hpp"
 
-Boat::Boat(Renderer const& r, int const& x, int const& y, int const& z) : Object(r, new Surface("Boat.png"), x, y, z) {
+Boat::Boat(Renderer const& r, int const& x, int const& y, float const& z) : Object(r, new Surface("Boat.png"), x, y, z) {
   //we'll need to set back the boat image;
 }
